Remove `include/grpcpp/impl/codegen/call.h`
