+++ include/grpcpp/impl/codegen/channel_interface.h
@@ -22,7 +22,7 @@
 // IWYU pragma: private
 
 #include <grpc/impl/codegen/connectivity_state.h>
-#include <grpcpp/impl/codegen/call.h>
+#include <grpcpp/impl/call.h>
 #include <grpcpp/impl/codegen/status.h>
 #include <grpcpp/impl/codegen/time.h>
