+++ include/grpcpp/impl/codegen/sync_stream.h
@@ -20,7 +20,7 @@
 
 // IWYU pragma: private, include <grpcpp/support/sync_stream.h>
 
-#include <grpcpp/impl/codegen/call.h>
+#include <grpcpp/impl/call.h>
 #include <grpcpp/impl/codegen/channel_interface.h>
 #include <grpcpp/impl/codegen/client_context.h>
 #include <grpcpp/impl/codegen/completion_queue.h>
