+++ include/grpcpp/test/mock_stream.h
@@ -23,7 +23,7 @@
 
 #include <gmock/gmock.h>
 
-#include <grpcpp/impl/codegen/call.h>
+#include <grpcpp/impl/call.h>
 #include <grpcpp/support/async_stream.h>
 #include <grpcpp/support/async_unary_call.h>
 #include <grpcpp/support/sync_stream.h>
